+++ FunkosShopBack-end/FunkosShopBack-end/Models/Entities/Producto.cs
@@ -4,7 +4,7 @@
     {
         public int ProductoID { get; set; }
         public string NombreProducto { get; set; }
-        public double PrecioEUR { get; set; }
+        public decimal PrecioEUR { get; set; }
         public string Descripcion { get; set; }
         public string Categoria { get; set; }
         public int Stock { get; set; }
